Add support for building snapshot images
This diff is a start at building snapshot images. Right now, the main
differences between the release workflow and the snapshot workflow is that:

* snapshots don't use signify
* snapshots make an additional http call to figure out what version suffix to
  use

There are potential problem if someone has a leftover mirror of old snapshots,
and reruns the script. This will try and continue downloading the set, but in
the case of snapshots, sets with the same name will have different lengths at
different times. If a user doesn't clean up their snapshots images, this can
lead to corruption. However, keeping them around is useful for multiple reruns
in the short-term, where this won't likely be an issue. A note in the README has
been added to address this.